use ints instead

+++ soh/soh/Enhancements/randomizer.h
@@ -12,7 +12,7 @@ class Randomizer {
     std::unordered_map<RandomizerCheck, RandomizerGet> itemLocations;
     GetItemID GetItemFromGet(RandomizerGet randoGet, GetItemID ogItemId);
     RandomizerCheck GetCheckFromActor(s16 actorId, GetItemID ogItemId);
-    RandomizerCheck GetCheckFromSceneAndParams(s16 sceneNum, s16 actorParams, f32 homePosX = 0, f32 homePosY = 0, f32 homePosZ = 0);
+    RandomizerCheck GetCheckFromSceneAndParams(s16 sceneNum, s16 actorParams, s32 homePosX = 0, s32 homePosY = 0, s32 homePosZ = 0);
 
   public:
     Randomizer();
@@ -22,7 +22,7 @@ class Randomizer {
     void ParseItemLocations(std::string spoilerfilename);
     GetItemID GetItemFromActor(s16 actorId, GetItemID ogItemId);
     GetItemID GetItemFromSceneAndParams(s16 sceneNum, s16 actorParams, GetItemID ogItemId);
-    GetItemID GetItemFromSceneParamsAndHomePos(s16 sceneNum, s16 actorParams, f32 homePosX, f32 homePosY, f32 homePosZ, GetItemID ogItemId);
+    GetItemID GetItemFromSceneParamsAndHomePos(s16 sceneNum, s16 actorParams, s32 homePosX, s32 homePosY, s32 homePosZ, GetItemID ogItemId);
 };
 
 #endif
